Added code for a simple UDP Transciever
QT -based C++ UDP Transceiver with a GUI. Super simple - designed as a good place to start if you are interested in sending or receiving messages via UDP. Be sure to do your research on the throughput of UDP connections - if you need more than 10 kb/s you probably are looking for a TCP Transciever - these are capable of much higher throughput around at least 64 kb/s. This project was made with QT Creator 4 so it should run on most devices.